pyproject.toml
[project]
name = "metamindbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "psycopg-binary (>=3.2.4,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "aiogram (==2.25.2)",
    "openai (==0.28)",
    "httpx (>=0.28.1,<0.29.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
